software/projects/django-bugnet/requirements.txt
Django>=1.4
django-macaddress
django-extensions
pyserial
